[Fix] Add create user test, Group test into versions
